+++ message.go
@@ -2,15 +2,9 @@ package dogma
 
 // A Message is an application-defined unit of data that describes a [Command],
 // [Event], or [Timeout] within a message-based application.
-//
-// Deprecated: Application developers should use [Command], [Event], or
-// [Timeout] instead.
 type Message interface {
 	// MessageDescription returns a human-readable description of the message.
 	MessageDescription() string
-
-	// Validate returns a non-nil error if the message is invalid.
-	Validate() error
 }
 
 // A Command is a message that represents a request for a Dogma application to
@@ -20,7 +14,7 @@ type Command interface {
 	MessageDescription() string
 
 	// Validate returns a non-nil error if the message is invalid.
-	Validate() error
+	Validate(CommandValidationScope) error
 }
 
 // An Event is a message that indicates that some action has occurred within a
@@ -30,7 +24,7 @@ type Event interface {
 	MessageDescription() string
 
 	// Validate returns a non-nil error if the message is invalid.
-	Validate() error
+	Validate(EventValidationScope) error
 }
 
 // A Timeout is a message that represents a request for an action to be
@@ -40,7 +34,7 @@ type Timeout interface {
 	MessageDescription() string
 
 	// Validate returns a non-nil error if the message is invalid.
-	Validate() error
+	Validate(TimeoutValidationScope) error
 }
 
 // UnexpectedMessage is a panic value used by a message handler when it receives
@@ -48,3 +42,21 @@ type Timeout interface {
 var UnexpectedMessage unexpectedMessage
 
 type unexpectedMessage struct{}
+
+// CommandValidationScope provides information about the context in which a
+// [Command] is being validated.
+type CommandValidationScope interface {
+	reservedCommandValidationScope()
+}
+
+// EventValidationScope provides information about the context in which an
+// [Event] is being validated.
+type EventValidationScope interface {
+	reservedEventValidationScope()
+}
+
+// TimeoutValidationScope provides information about the context in which a
+// [Timeout] is being validated.
+type TimeoutValidationScope interface {
+	reservedTimeoutValidationScope()
+}